Refactor: Move plan schedulers to plan_schedulers dir, make it more cleaner
